Ball Like,DisLike,Cancel Action Service 구분 Reply 에 평가 결과 연결

+++ src/main/java/com/wing/forutona/ForutonaUser/Dto/FUserInfoResDto.java
@@ -14,7 +14,7 @@ public class FUserInfoResDto {
     private String uid;
     private String nickName;
     private String profilePictureUrl;
-    private Long gender;
+    private Integer gender;
     private LocalDate ageDate;
     private String email;
     private Boolean forutonaAgree;
